requirements.txt
fastapi
uvicorn[standard]
gunicorn
httpx
loguru
sqlalchemy
psycopg2-binary
boto3
requests
python-multipart
pathlib
prometheus-client
PyJWT
cryptography
pyjwt[crypto]